Furnace is now nameable

+++ src/main/java/cn/nukkit/Server.java
@@ -1108,8 +1108,8 @@ public class Server {
         return this.getPropertyBoolean("generate-structures", true);
     }
 
-    public byte getGamemode() {
-        return (byte) (this.getPropertyInt("gamemode", 0) & 0b11);
+    public int getGamemode() {
+        return this.getPropertyInt("gamemode", 0) & 0b11;
     }
 
     public boolean getForceGamemode() {
